refactor(#122): split core/qtl_constants.py into per-plugin and per-algorithm homes
- EQTL_SOURCES -> skills/gtex_eqtl/shared/constants.py
- PQTL_SOURCES, FANG_TISSUES, FANG_TISSUE_EQTL_MAPPING -> skills/pqtl/shared/constants.py
- CASCADE_*, DEFAULT_COLOC_*, DEFAULT_{EQTL,PQTL}_P_THRESHOLD
  -> algorithms/qtlcascade/constants.py (NEW).
- Removed the dead skill-bound re-exports (EQTL_SOURCES, PQTL_SOURCES,
  FANG_*) from algorithms/qtlcascade/__init__.py (no external consumer;
  re-pointing them to skills/ would violate the algorithms->skills rule).
- Deleted the now-empty core/qtl_constants.py.

One of several PRs closing #122.
