+++ hvantk/algorithms/qtlcascade/coloc.py
@@ -29,7 +29,7 @@ import numpy as np
 import pandas as pd
 
 from hvantk.core.models.backends import Backend, algorithm
-from hvantk.core.qtl_constants import (
+from hvantk.algorithms.qtlcascade.constants import (
     DEFAULT_COLOC_P1,
     DEFAULT_COLOC_P2,
     DEFAULT_COLOC_P12,
