+++ hvantk/algorithms/qtlcascade/constants.py
@@ -1,5 +1,6 @@
-"""
-Constants and default parameters for QTL cascade analysis.
+"""Algorithm-local constants for QTL cascade analysis.
+
+Moved from hvantk/core/qtl_constants.py per issue #122 (clean-core principle).
 
 References
 ----------
@@ -17,7 +18,7 @@ CASCADE_CLASSES = ("eqtl_mediated", "discordant", "eqtl_only", "pqtl_only")
 
 CASCADE_CLASS_LABELS = {
     "eqtl_mediated": "eQTL-mediated pQTL",
-    "discordant": "Discordant (eQTL \u2260 pQTL direction)",
+    "discordant": "Discordant (eQTL ≠ pQTL direction)",
     "eqtl_only": "eQTL only (no pQTL)",
     "pqtl_only": "pQTL only (no eQTL)",
 }
@@ -53,27 +54,3 @@ DEFAULT_COLOC_H4_THRESHOLD = 0.8
 
 # Regional window for coloc (±kb from lead variant)
 DEFAULT_COLOC_WINDOW_KB = 500
-
-# ---------------------------------------------------------------------------
-# Fang et al. (2025) tissue-matched pQTL tissues
-# ---------------------------------------------------------------------------
-
-FANG_TISSUES = ("Colon", "Heart", "Liver", "Lung", "Thyroid")
-
-# Mapping from Fang pQTL tissue names to GTEx eQTL sub-tissue names.
-# Some pQTL tissues span multiple eQTL sub-tissues.
-FANG_TISSUE_EQTL_MAPPING = {
-    "Colon": ("Colon_Transverse", "Colon_Sigmoid"),
-    "Heart": ("Heart_Left_Ventricle", "Heart_Atrial_Appendage"),
-    "Liver": ("Liver",),
-    "Lung": ("Lung",),
-    "Thyroid": ("Thyroid",),
-}
-
-# ---------------------------------------------------------------------------
-# Supported data source identifiers
-# ---------------------------------------------------------------------------
-
-EQTL_SOURCES = ("gtex_v11", "gtex_v8", "eqtlgen")
-# Only gtex_fang is currently implemented; others are planned.
-PQTL_SOURCES = ("gtex_fang",)
