+++ hvantk/skills/gtex_eqtl/shared/constants.py
@@ -1,0 +1,10 @@
+"""Plugin-local constants for the gtex_eqtl skill.
+
+Moved from hvantk/core/qtl_constants.py per issue #122 (clean-core principle).
+"""
+
+# ---------------------------------------------------------------------------
+# Supported data source identifiers
+# ---------------------------------------------------------------------------
+
+EQTL_SOURCES = ("gtex_v11", "gtex_v8", "eqtlgen")
